flask-server/requirements.txt
blinker==1.7.0
boto3==1.34.84
botocore==1.34.84
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.1
Flask-Cors==4.0.0
flask-sqlalchemy==3.1.1
GDAL==3.4.2
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.22.3
protobuf==3.19.4
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.29
typing-extensions==4.11.0
urllib3==1.26.18
werkzeug==3.0.1
zipp==3.17.0
